dcp-815 separate class for metadata/service.py
