+++ exporter.py
@@ -8,8 +8,8 @@ from kombu import Connection
 
 from exporter.amqp import AmqpConnConfig, QueueConfig
 from exporter.graph.graph_crawler import GraphCrawler
+from exporter.metadata.service import MetadataService
 from exporter.schema import SchemaService
-from exporter.scratch import MetadataService
 from exporter.session_context import configure_logger
 from exporter.terra.dcp_staging_client import DcpStagingClient
 from exporter.terra.terra_export_job import TerraExportJobService
