+++ exporter/graph/graph_crawler.py
@@ -7,7 +7,7 @@ from typing import List, Iterable, Optional, Callable
 from exporter.graph.experiment_graph import ExperimentGraph, ProcessLink, Input, Output, \
     ProtocolLink, SupplementaryFileLink, SupplementedEntity, SupplementaryFile
 from exporter.metadata.resource import MetadataResource
-from exporter.scratch import MetadataService
+from exporter.metadata.service import MetadataService
 
 
 @dataclass
