+++ exporter/terra/terra_exporter.py
@@ -5,7 +5,7 @@ from hca_ingest.api.ingestapi import IngestApi
 from exporter import utils
 from exporter.graph.graph_crawler import GraphCrawler
 from exporter.metadata.resource import MetadataResource
-from exporter.scratch import MetadataService
+from exporter.metadata.service import MetadataService
 from exporter.terra.dcp_staging_client import DcpStagingClient
 from exporter.terra.terra_export_job import TerraExportJobService
