+++ tests/exporter/graph/test_graph_crawler.py
@@ -5,7 +5,7 @@ from mock import MagicMock
 
 from exporter.graph.graph_crawler import GraphCrawler
 from exporter.metadata.resource import MetadataResource
-from exporter.scratch import MetadataService
+from exporter.metadata.service import MetadataService
 from tests.mocks.files import MockEntityFiles
 from tests.mocks.ingest import MockIngestAPI
