+++ tests/exporter/terra/test_dcp_staging_client.py
@@ -6,8 +6,8 @@ from hca_ingest.api.ingestapi import IngestApi
 from exporter.graph.graph_crawler import GraphCrawler
 from exporter.metadata.exceptions import MetadataParseException
 from exporter.metadata.resource import MetadataResource
+from exporter.metadata.service import MetadataService
 from exporter.schema import SchemaService
-from exporter.scratch import MetadataService
 from exporter.terra.dcp_staging_client import DcpStagingClient
 from tests.mocks.files import MockEntityFiles
